+++ Mime_extensions_test.v
@@ -1,0 +1,7 @@
+module test
+
+import mime { Mime }
+
+fn test_returns_html_extensions() {
+    assert Mime.text_html.extensions() == ["html", "htm"]
+}