+++ Mime_str.v
@@ -1,0 +1,8 @@
+module mime
+
+pub fn (this Mime) str() string {
+    match this {
+        .text_html { return "text/html" }
+        .application_json { return "application/json" }
+    }
+}